pyproject.toml
[project]
name = "ai-doc-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.9.0",
    "dotenv-python>=0.0.1",
    "fastapi>=0.118.0",
    "isort>=6.1.0",
    "jinja2>=3.1.6",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-ollama>=0.3.10",
    "langchain-pinecone>=0.2.12",
    "langchain-tavily>=0.2.11",
    "nltk>=3.9.2",
    "pinecone[asyncio]>=7.3.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
    "streamlit-chat>=0.1.1",
    "tqdm>=4.67.1",
]
